level1->part5->AdminFacade

+++ src/DBDAO/CompaniesDBDAO.java
@@ -3,6 +3,7 @@ package DBDAO;
 import CLS.*;
 import DAO.*;
 import JavaBeans.Company;
+import JavaBeans.Coupon;
 
 import java.sql.*;
 import java.util.ArrayList;
@@ -34,19 +35,25 @@ public class CompaniesDBDAO implements CompaniesDAO {
     }
 
     @Override
-    public void addCompany(Company company) {
-        final String ADD_COMPANY = "INSERT INTO `jb_project`.`companies` (`id`, `NAME`, `EMAIL`, `PASSWORD`) VALUES (?,?,?,?);";
-        Map<Integer, Object> params = new HashMap<>();
-        params.put(1, company.getID());
-        params.put(2, company.getName());
-        params.put(3, company.getEmail());
-        params.put(4, company.getPassword());
+    public void addCompany(Company company) throws SQLException {
 
-        boolean flag = DBtools.runQuery(ADD_COMPANY, params);
-        if (flag)
-            System.out.println("The Company has been added successfully.");
-        else
-            System.out.println("ERROR :The Company not added.");
+        if (checkNameIsExists(company.getName())) {
+            System.out.println(" company name already exists.");
+        } else if (checkEmailIsExists(company.getEmail())) {
+            System.out.println(" company email already exists.");
+        } else {
+            final String ADD_COMPANY = "INSERT INTO `jb_project`.`companies` (`id`, `NAME`, `EMAIL`, `PASSWORD`) VALUES (?,?,?,?);";
+            Map<Integer, Object> params = new HashMap<>();
+            params.put(1, company.getID());
+            params.put(2, company.getName());
+            params.put(3, company.getEmail());
+            params.put(4, company.getPassword());
+            boolean flag = DBtools.runQuery(ADD_COMPANY, params);
+            if (flag)
+                System.out.println("The Company has been added successfully.");
+            else
+                System.out.println("ERROR :The Company not added.");
+        }
 
 
     }
@@ -54,16 +61,24 @@ public class CompaniesDBDAO implements CompaniesDAO {
     @Override
     public void updateCompany(Company company) throws SQLException {
 
-        if (isCompanyExists(company.getEmail(), company.getPassword())) {
-            final String id = String.valueOf(company.getID());
-            final String UPDATE = "UPDATE `jb_project`.`companies` SET `id` = ?, `NAME` = ?, `EMAIL` = ?, `PASSWORD` = ? WHERE (`id` = " + id + ");";
+
+        if (checkIdIsExists(company.getID()) && checkNameIsExists(company.getName())) {
+            final int id = company.getID();
+            final String name = company.getName();
+            final String UPDATE = "UPDATE `jb_project`.`companies` SET `EMAIL` = ?, `PASSWORD` = ? WHERE `id` = "+id+";";
             Map<Integer, Object> params = new HashMap<>();
-            params.put(1, company.getID());
-            params.put(2, company.getName());
-            params.put(3, company.getEmail());
-            params.put(4, company.getPassword());
-            DBtools.runQuery(UPDATE, params);
-            System.out.println("The Company has been Updated successfully.");
+
+            params.put(1, company.getEmail());
+            params.put(2, company.getPassword());
+
+
+            boolean flag = DBtools.runQuery(UPDATE, params);
+            if (flag) {
+                System.out.println("The Company has been Updated successfully.");
+            }else {
+                System.out.println("Field Updated company.");
+
+            }
         } else {
             System.out.println("The Company not exists.");
         }
@@ -72,13 +87,19 @@ public class CompaniesDBDAO implements CompaniesDAO {
     }
 
     @Override
-    public void deleteCompany(int companyID) {
-        final String DELETE = "DELETE FROM `jb_project`.`companies` WHERE id = " + companyID + ";";
-        boolean flag = DBtools.runQuery(DELETE);
-        if (flag)
-            System.out.println("The Company deleted successfully.");
-        else
-            System.out.println("ERROR ,the company not deleted.");
+    public void deleteCompany(int companyID) throws SQLException {
+        if(checkIdIsExists(companyID)) {
+            final String DELETE_COMPANY = "DELETE FROM `jb_project`.`companies` WHERE id = " + companyID + ";";
+            final String DELETE_COUPONS = "DELETE FROM `jb_project`.`coupons` WHERE COMPANY_ID = " + companyID + ";";
+            boolean deleteCoupons = DBtools.runQuery(DELETE_COUPONS);
+            boolean deleteCompany = DBtools.runQuery(DELETE_COMPANY);
+            if (deleteCompany && deleteCoupons)
+                System.out.println("The Company deleted successfully.");
+            else
+                System.out.println("ERROR ,the company not deleted.");
+        }else {
+            System.out.println("The Company not exists.");
+        }
 
     }
 
@@ -93,7 +114,9 @@ public class CompaniesDBDAO implements CompaniesDAO {
             String name = resultSet.getString(2);
             String email = resultSet.getString(3);
             String pass = resultSet.getString(4);
-            companies.add(new Company(ID, name, email, pass, new ArrayList<>()));
+            ArrayList<Coupon> coupons = (new CouponsDBDAO()).getAllCouponsWithSameCompanyID(ID);
+
+            companies.add(new Company(ID, name, email, pass, coupons));
         }
         return companies;
 
@@ -101,18 +124,69 @@ public class CompaniesDBDAO implements CompaniesDAO {
 
     @Override
     public Company getOneCompany(int companyID) throws SQLException {
-        final String ONE_COMPANIES = "SELECT * FROM `jb_project`.`companies` WHERE id = " + companyID + ";";
-        Map<Integer, Object> params = new HashMap<>();
-        ResultSet resultSet = DBtools.runQueryForResult(ONE_COMPANIES, params);
-        while (resultSet.next()) {
+        if (checkIdIsExists(companyID)) {
+            final String ONE_COMPANIES = "SELECT * FROM `jb_project`.`companies` WHERE id = " + companyID + ";";
+            Map<Integer, Object> params = new HashMap<>();
+            ResultSet resultSet = DBtools.runQueryForResult(ONE_COMPANIES, params);
+            while (resultSet.next()) {
 
-            int ID = resultSet.getInt(1);
-            String name = resultSet.getString(2);
-            String email = resultSet.getString(3);
-            String pass = resultSet.getString(4);
-            return new Company(ID, name, email, pass, new ArrayList<>());
+                int ID = resultSet.getInt(1);
+                String name = resultSet.getString(2);
+                String email = resultSet.getString(3);
+                String pass = resultSet.getString(4);
+                ArrayList<Coupon> coupons = (new CouponsDBDAO()).getAllCouponsWithSameCompanyID(ID);
+                return new Company(ID, name, email, pass, coupons);
+            }
+            return null;
+        } else {
+            System.out.println("The Company not exists.");
+            return null;
         }
-        return null;
+    }
+
+    public boolean checkNameIsExists(String name) throws SQLException {
+        // check if the Name o is EXISTS;
+        boolean checkName = false;
+        final String CHECK = "SELECT * FROM `jb_project`.`companies`;";
+        Map<Integer, Object> params = new HashMap<>();
+        ResultSet result = DBtools.runQueryForResult(CHECK, params);
+        while (result.next()) {
+            String name_ = result.getNString(2);
+            if (name.equals(name_))
+                return true;
+        }
+        return false;
+
+    }
+
+    public boolean checkIdIsExists(int id) throws SQLException {
+        // check if the ID o is EXISTS;
+        boolean checkID = false;
+        final String CHECK = "SELECT * FROM `jb_project`.`companies`;";
+        Map<Integer, Object> params = new HashMap<>();
+        ResultSet result = DBtools.runQueryForResult(CHECK, params);
+        while (result.next()) {
+            int id_ = result.getInt(1);
+            if (id == id_)
+                return true;
+        }
+        return false;
+
+    }
+
+    public boolean checkEmailIsExists(String email) throws SQLException {
+        // check if the Email o is EXISTS;
+        boolean checkEmail = false;
+        final String CHECK = "SELECT * FROM `jb_project`.`companies`;";
+        Map<Integer, Object> params = new HashMap<>();
+        ResultSet result = DBtools.runQueryForResult(CHECK, params);
+        while (result.next()) {
+            String email_ = result.getNString(3);
+            if (email.equals(email_))
+                return true;
+        }
+        return false;
+
     }
